+++ sdk/appservice/arm-appservice/src/models/index.ts
@@ -4243,6 +4243,70 @@ export interface WebJobCollection {
   readonly nextLink?: string;
 }
 
+/** The workflow filter. */
+export interface WorkflowArtifacts {
+  /** Application settings of the workflow. */
+  appSettings?: Record<string, unknown>;
+  /** Files of the app. */
+  files?: { [propertyName: string]: Record<string, unknown> };
+  /** Files of the app to delete. */
+  filesToDelete?: string[];
+}
+
+/** Collection of Kudu workflow information elements. */
+export interface WorkflowEnvelopeCollection {
+  /** Collection of resources. */
+  value: WorkflowEnvelope[];
+  /**
+   * Link to next page of resources.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly nextLink?: string;
+}
+
+/** Workflow properties definition. */
+export interface WorkflowEnvelope {
+  /**
+   * The resource id.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly id?: string;
+  /**
+   * Gets the resource name.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly name?: string;
+  /**
+   * Gets the resource type.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly type?: string;
+  /** The resource kind. */
+  kind?: string;
+  /** The resource location. */
+  location?: string;
+  /** Additional workflow properties. */
+  properties?: WorkflowEnvelopeProperties;
+}
+
+/** Additional workflow properties. */
+export interface WorkflowEnvelopeProperties {
+  /** Gets or sets the files. */
+  files?: { [propertyName: string]: Record<string, unknown> };
+  /** Gets or sets the state of the workflow. */
+  flowState?: WorkflowState;
+  /** Gets or sets workflow health. */
+  health?: WorkflowHealth;
+}
+
+/** Represents the workflow health. */
+export interface WorkflowHealth {
+  /** Gets or sets the workflow health state. */
+  state: WorkflowHealthState;
+  /** Gets or sets the workflow error. */
+  error?: ErrorEntity;
+}
+
 /** The access key regenerate action content. */
 export interface RegenerateActionParameter {
   /** The key type. */
@@ -9363,8 +9427,8 @@ export enum KnownRevisionProvisioningState {
  */
 export type RevisionProvisioningState = string;
 
-/** Known values of {@link Enum14} that the service accepts. */
-export enum KnownEnum14 {
+/** Known values of {@link ProviderOsTypeSelected} that the service accepts. */
+export enum KnownProviderOsTypeSelected {
   /** Windows */
   Windows = "Windows",
   /** Linux */
@@ -9378,8 +9442,8 @@ export enum KnownEnum14 {
 }
 
 /**
- * Defines values for Enum14. \
- * {@link KnownEnum14} can be used interchangeably with Enum14,
+ * Defines values for ProviderOsTypeSelected. \
+ * {@link KnownProviderOsTypeSelected} can be used interchangeably with ProviderOsTypeSelected,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
  * **Windows** \
@@ -9388,10 +9452,10 @@ export enum KnownEnum14 {
  * **LinuxFunctions** \
  * **All**
  */
-export type Enum14 = string;
+export type ProviderOsTypeSelected = string;
 
-/** Known values of {@link Enum15} that the service accepts. */
-export enum KnownEnum15 {
+/** Known values of {@link ProviderStackOsType} that the service accepts. */
+export enum KnownProviderStackOsType {
   /** Windows */
   Windows = "Windows",
   /** Linux */
@@ -9401,105 +9465,15 @@ export enum KnownEnum15 {
 }
 
 /**
- * Defines values for Enum15. \
- * {@link KnownEnum15} can be used interchangeably with Enum15,
+ * Defines values for ProviderStackOsType. \
+ * {@link KnownProviderStackOsType} can be used interchangeably with ProviderStackOsType,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
  * **Windows** \
  * **Linux** \
  * **All**
  */
-export type Enum15 = string;
-
-/** Known values of {@link Enum16} that the service accepts. */
-export enum KnownEnum16 {
-  /** Windows */
-  Windows = "Windows",
-  /** Linux */
-  Linux = "Linux",
-  /** All */
-  All = "All"
-}
-
-/**
- * Defines values for Enum16. \
- * {@link KnownEnum16} can be used interchangeably with Enum16,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **Windows** \
- * **Linux** \
- * **All**
- */
-export type Enum16 = string;
-
-/** Known values of {@link Enum17} that the service accepts. */
-export enum KnownEnum17 {
-  /** Windows */
-  Windows = "Windows",
-  /** Linux */
-  Linux = "Linux",
-  /** All */
-  All = "All"
-}
-
-/**
- * Defines values for Enum17. \
- * {@link KnownEnum17} can be used interchangeably with Enum17,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **Windows** \
- * **Linux** \
- * **All**
- */
-export type Enum17 = string;
-
-/** Known values of {@link Enum18} that the service accepts. */
-export enum KnownEnum18 {
-  /** Windows */
-  Windows = "Windows",
-  /** Linux */
-  Linux = "Linux",
-  /** All */
-  All = "All"
-}
-
-/**
- * Defines values for Enum18. \
- * {@link KnownEnum18} can be used interchangeably with Enum18,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **Windows** \
- * **Linux** \
- * **All**
- */
-export type Enum18 = string;
-
-/** Known values of {@link Enum19} that the service accepts. */
-export enum KnownEnum19 {
-  /** Windows */
-  Windows = "Windows",
-  /** Linux */
-  Linux = "Linux",
-  /** WindowsFunctions */
-  WindowsFunctions = "WindowsFunctions",
-  /** LinuxFunctions */
-  LinuxFunctions = "LinuxFunctions",
-  /** All */
-  All = "All"
-}
-
-/**
- * Defines values for Enum19. \
- * {@link KnownEnum19} can be used interchangeably with Enum19,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **Windows** \
- * **Linux** \
- * **WindowsFunctions** \
- * **LinuxFunctions** \
- * **All**
- */
-export type Enum19 = string;
+export type ProviderStackOsType = string;
 
 /** Known values of {@link ResourceScopeType} that the service accepts. */
 export enum KnownResourceScopeType {
@@ -10028,36 +10002,6 @@ export enum KnownWorkflowTriggerProvisioningState {
  * **Completed**
  */
 export type WorkflowTriggerProvisioningState = string;
-
-/** Known values of {@link WorkflowState} that the service accepts. */
-export enum KnownWorkflowState {
-  /** NotSpecified */
-  NotSpecified = "NotSpecified",
-  /** Completed */
-  Completed = "Completed",
-  /** Enabled */
-  Enabled = "Enabled",
-  /** Disabled */
-  Disabled = "Disabled",
-  /** Deleted */
-  Deleted = "Deleted",
-  /** Suspended */
-  Suspended = "Suspended"
-}
-
-/**
- * Defines values for WorkflowState. \
- * {@link KnownWorkflowState} can be used interchangeably with WorkflowState,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **NotSpecified** \
- * **Completed** \
- * **Enabled** \
- * **Disabled** \
- * **Deleted** \
- * **Suspended**
- */
-export type WorkflowState = string;
 
 /** Known values of {@link RecurrenceFrequency} that the service accepts. */
 export enum KnownRecurrenceFrequency {
@@ -10604,6 +10548,20 @@ export type PublicCertificateLocation =
 export type SiteExtensionType = "Gallery" | "WebRoot";
 /** Defines values for TriggeredWebJobStatus. */
 export type TriggeredWebJobStatus = "Success" | "Failed" | "Error";
+/** Defines values for WorkflowState. */
+export type WorkflowState =
+  | "NotSpecified"
+  | "Completed"
+  | "Enabled"
+  | "Disabled"
+  | "Deleted"
+  | "Suspended";
+/** Defines values for WorkflowHealthState. */
+export type WorkflowHealthState =
+  | "NotSpecified"
+  | "Healthy"
+  | "Unhealthy"
+  | "Unknown";
 /** Defines values for DaysOfWeek. */
 export type DaysOfWeek =
   | "Sunday"
@@ -12343,7 +12301,7 @@ export type KubeEnvironmentsListByResourceGroupNextResponse = KubeEnvironmentCol
 /** Optional parameters. */
 export interface ProviderGetAvailableStacksOptionalParams
   extends coreClient.OperationOptions {
-  osTypeSelected?: Enum14;
+  osTypeSelected?: ProviderOsTypeSelected;
 }
 
 /** Contains response data for the getAvailableStacks operation. */
@@ -12353,7 +12311,7 @@ export type ProviderGetAvailableStacksResponse = ApplicationStackCollection;
 export interface ProviderGetFunctionAppStacksOptionalParams
   extends coreClient.OperationOptions {
   /** Stack OS Type */
-  stackOsType?: Enum15;
+  stackOsType?: ProviderStackOsType;
 }
 
 /** Contains response data for the getFunctionAppStacks operation. */
@@ -12363,7 +12321,7 @@ export type ProviderGetFunctionAppStacksResponse = FunctionAppStackCollection;
 export interface ProviderGetFunctionAppStacksForLocationOptionalParams
   extends coreClient.OperationOptions {
   /** Stack OS Type */
-  stackOsType?: Enum16;
+  stackOsType?: ProviderStackOsType;
 }
 
 /** Contains response data for the getFunctionAppStacksForLocation operation. */
@@ -12373,7 +12331,7 @@ export type ProviderGetFunctionAppStacksForLocationResponse = FunctionAppStackCo
 export interface ProviderGetWebAppStacksForLocationOptionalParams
   extends coreClient.OperationOptions {
   /** Stack OS Type */
-  stackOsType?: Enum17;
+  stackOsType?: ProviderStackOsType;
 }
 
 /** Contains response data for the getWebAppStacksForLocation operation. */
@@ -12390,7 +12348,7 @@ export type ProviderListOperationsResponse = CsmOperationCollection;
 export interface ProviderGetWebAppStacksOptionalParams
   extends coreClient.OperationOptions {
   /** Stack OS Type */
-  stackOsType?: Enum18;
+  stackOsType?: ProviderStackOsType;
 }
 
 /** Contains response data for the getWebAppStacks operation. */
@@ -12399,7 +12357,7 @@ export type ProviderGetWebAppStacksResponse = WebAppStackCollection;
 /** Optional parameters. */
 export interface ProviderGetAvailableStacksOnPremOptionalParams
   extends coreClient.OperationOptions {
-  osTypeSelected?: Enum19;
+  osTypeSelected?: ProviderOsTypeSelected;
 }
 
 /** Contains response data for the getAvailableStacksOnPrem operation. */
@@ -12408,7 +12366,7 @@ export type ProviderGetAvailableStacksOnPremResponse = ApplicationStackCollectio
 /** Optional parameters. */
 export interface ProviderGetAvailableStacksNextOptionalParams
   extends coreClient.OperationOptions {
-  osTypeSelected?: Enum14;
+  osTypeSelected?: ProviderOsTypeSelected;
 }
 
 /** Contains response data for the getAvailableStacksNext operation. */
@@ -12418,7 +12376,7 @@ export type ProviderGetAvailableStacksNextResponse = ApplicationStackCollection;
 export interface ProviderGetFunctionAppStacksNextOptionalParams
   extends coreClient.OperationOptions {
   /** Stack OS Type */
-  stackOsType?: Enum15;
+  stackOsType?: ProviderStackOsType;
 }
 
 /** Contains response data for the getFunctionAppStacksNext operation. */
@@ -12428,7 +12386,7 @@ export type ProviderGetFunctionAppStacksNextResponse = FunctionAppStackCollectio
 export interface ProviderGetFunctionAppStacksForLocationNextOptionalParams
   extends coreClient.OperationOptions {
   /** Stack OS Type */
-  stackOsType?: Enum16;
+  stackOsType?: ProviderStackOsType;
 }
 
 /** Contains response data for the getFunctionAppStacksForLocationNext operation. */
@@ -12438,7 +12396,7 @@ export type ProviderGetFunctionAppStacksForLocationNextResponse = FunctionAppSta
 export interface ProviderGetWebAppStacksForLocationNextOptionalParams
   extends coreClient.OperationOptions {
   /** Stack OS Type */
-  stackOsType?: Enum17;
+  stackOsType?: ProviderStackOsType;
 }
 
 /** Contains response data for the getWebAppStacksForLocationNext operation. */
@@ -12455,7 +12413,7 @@ export type ProviderListOperationsNextResponse = CsmOperationCollection;
 export interface ProviderGetWebAppStacksNextOptionalParams
   extends coreClient.OperationOptions {
   /** Stack OS Type */
-  stackOsType?: Enum18;
+  stackOsType?: ProviderStackOsType;
 }
 
 /** Contains response data for the getWebAppStacksNext operation. */
@@ -12464,7 +12422,7 @@ export type ProviderGetWebAppStacksNextResponse = WebAppStackCollection;
 /** Optional parameters. */
 export interface ProviderGetAvailableStacksOnPremNextOptionalParams
   extends coreClient.OperationOptions {
-  osTypeSelected?: Enum19;
+  osTypeSelected?: ProviderOsTypeSelected;
 }
 
 /** Contains response data for the getAvailableStacksOnPremNext operation. */
@@ -16420,6 +16378,62 @@ export interface WebAppsGetWebJobOptionalParams
 export type WebAppsGetWebJobResponse = WebJob;
 
 /** Optional parameters. */
+export interface WebAppsDeployWorkflowArtifactsOptionalParams
+  extends coreClient.OperationOptions {
+  /** Application settings and files of the workflow. */
+  workflowArtifacts?: WorkflowArtifacts;
+}
+
+/** Optional parameters. */
+export interface WebAppsDeployWorkflowArtifactsSlotOptionalParams
+  extends coreClient.OperationOptions {
+  /** Application settings and files of the workflow. */
+  workflowArtifacts?: WorkflowArtifacts;
+}
+
+/** Optional parameters. */
+export interface WebAppsListInstanceWorkflowsSlotOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listInstanceWorkflowsSlot operation. */
+export type WebAppsListInstanceWorkflowsSlotResponse = WorkflowEnvelopeCollection;
+
+/** Optional parameters. */
+export interface WebAppsGetInstanceWorkflowSlotOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the getInstanceWorkflowSlot operation. */
+export type WebAppsGetInstanceWorkflowSlotResponse = WorkflowEnvelope;
+
+/** Optional parameters. */
+export interface WebAppsListInstanceWorkflowsConfigurationConnectionsSlotOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listInstanceWorkflowsConfigurationConnectionsSlot operation. */
+export type WebAppsListInstanceWorkflowsConfigurationConnectionsSlotResponse = WorkflowEnvelope;
+
+/** Optional parameters. */
+export interface WebAppsListWorkflowsOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listWorkflows operation. */
+export type WebAppsListWorkflowsResponse = WorkflowEnvelopeCollection;
+
+/** Optional parameters. */
+export interface WebAppsGetWorkflowOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the getWorkflow operation. */
+export type WebAppsGetWorkflowResponse = WorkflowEnvelope;
+
+/** Optional parameters. */
+export interface WebAppsListWorkflowsConfigurationOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listWorkflowsConfiguration operation. */
+export type WebAppsListWorkflowsConfigurationResponse = WorkflowEnvelope;
+
+/** Optional parameters. */
 export interface WebAppsListNextOptionalParams
   extends coreClient.OperationOptions {}
 
@@ -16888,6 +16902,20 @@ export interface WebAppsListWebJobsNextOptionalParams
 
 /** Contains response data for the listWebJobsNext operation. */
 export type WebAppsListWebJobsNextResponse = WebJobCollection;
+
+/** Optional parameters. */
+export interface WebAppsListInstanceWorkflowsSlotNextOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listInstanceWorkflowsSlotNext operation. */
+export type WebAppsListInstanceWorkflowsSlotNextResponse = WorkflowEnvelopeCollection;
+
+/** Optional parameters. */
+export interface WebAppsListWorkflowsNextOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listWorkflowsNext operation. */
+export type WebAppsListWorkflowsNextResponse = WorkflowEnvelopeCollection;
 
 /** Optional parameters. */
 export interface WorkflowsRegenerateAccessKeyOptionalParams